using Logger

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task.status.model';
 import { AuthGuard } from '@nestjs/passport';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 @UseGuards(AuthGuard())
@@ -15,12 +16,12 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) {}
 
-    async getTaskFilter(filterTask: GetTasksFilterDto) :Promise<Task[]> {
-        return this.taskRepository.getTasks(filterTask);
+    async getTaskFilter(filterTask: GetTasksFilterDto, user: User) :Promise<Task[]> {
+        return this.taskRepository.getTasks(filterTask, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
         if(!found) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
@@ -28,8 +29,8 @@ export class TasksService {
     }
     
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
     // async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
@@ -42,17 +43,16 @@ export class TasksService {
     //     return task;
     // }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
 
         if(result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
-        
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
